Refactor remaining SDFG utilities to dace.sdfg.utils
