+++ dace/sdfg/__init__.py
@@ -1,30 +1,27 @@
 from dace.sdfg.sdfg import (
-    SDFG,  # sdfg.py
+    # sdfg.py
+    SDFG,
     InterstateEdge,
-    SDFGState,  # state.py
-    SubgraphView,  # graph.py
+    SDFGState,
 
-    # utils.py
-    has_dynamic_map_inputs,
-    dynamic_map_inputs,
-    is_parallel,
-    concurrent_subgraphs,
-    find_input_arraynode,
-    find_output_arraynode,
-    trace_nested_access,
-    is_array_stream_view,
-    local_transients,
+    # graph.py
+    SubgraphView,
 
-    # scopes.py
+    # scope.py
     ScopeSubgraphView,
-    is_devicelevel,
-    devicelevel_block_size,
 
     # replace.py
     replace,
     replace_properties)
 
-from dace.sdfg.scope import scope_contains_scope
+from dace.sdfg.utils import (has_dynamic_map_inputs, dynamic_map_inputs,
+                             is_parallel, concurrent_subgraphs,
+                             find_input_arraynode, find_output_arraynode,
+                             trace_nested_access, is_array_stream_view,
+                             local_transients)
+
+from dace.sdfg.scope import (scope_contains_scope, is_devicelevel,
+                             devicelevel_block_size)
 
 from dace.sdfg.validation import (InvalidSDFGError, InvalidSDFGNodeError,
                                   InvalidSDFGEdgeError,
